requirements/travis.txt
-r base.txt

factory_boy>=2.5
pytest
pytest-cov
pytest-django>=2.8,<2.9
pytest-capturelog
coveralls
MySQL-python
psycopg2
pep8
flake8
